ENCD-4855-wide-facets
Move internal_tags display component back to objectutils.js for another ticket that will add internal_tags displays to search results. Correct merge error for matrix visualizations. Return matrix visualizations to audit matrix.
